requirements.txt
pydantic==1.10.2
PyGithub==1.57
pytest==7.2.1
python-dotenv==0.21.1
PyYAML==6.0
requests==2.28.1
scikit_learn==1.2.1
